INBUS implementation moved to new package in common
